Update decode interface (#3476)
**What this PR does / why we need it**:

Previously, the decode interface is
```go
type ShardDecoder interface {
	// Decode unmarshals all given raw data parts to a given entity e.
	Decode(e interface{}, parts ...[]byte) error
}
```
But in some case, we need to know which shard these parts of data belongs to in order to merge (decode) those data correctly. This PR updates the decode interface to
```go
type ShardDecoder interface {
	// Decode unmarshals all given raw data parts to a given entity e.
	Decode(e interface{}, parts map[Shard][]byte) error
}
```

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.